Simplify desktop canvas
Add the eventHandlers type to the client so it can  be used instead of an
emitter and also instantiate the client in `useDesktopSession` instead
of the `useTdpClientCanvas`

Move client calls to `useDesktopSession`. `useTdpClientCanvas` shouldn't
set anything when it comes to client callbacks. Really, the only thing
`useTdpClientCanvas` should be is a glorified object of canvas callbacks
that we can spread as props into TdpClientCanvas.

From the looks of it, we still need it in `useDesktopSession` because of
the canvasOnPngFrame and canvasOnnBmpFrame methods that call a draw to
the canvas. We basically return those methods from `useTdpClientCanvas`
and then call them in the `useDesktopSession` `onPngFrame` etc.

Add typesafe eventHandler update method. We can set any number of handlers
in the constructor, but this method will allow us to set (or update) event
handlers during the lifetime of the client. It's also type safe so the
callsite knows exactly which handler keys can be added and their signature

Co-Authored-By: Michael Myers <michael.myers@goteleport.com>