Align cells across feature matrix tables
Make it easier for a user to follow a single column across multiple
tables of the feature matrix. Since the style sets cells to 25% width,
which assumes a table that is four cells wide, use a `style` tag so the
style only applies to the tables in the `teleport-editions.mdx` guide.

Closes gravitational/docs-website#134